move: 메인커뮤니티스타일 파일을 style - pages 폴더로 이동
