+++ src/styles/pages/MainCommunityStyle.tsx
